fix jwt auth; cicd lint stage added
