+++ .github/workflows/lint.yml
@@ -1,0 +1,21 @@
+name: Lint with Black
+
+on: [push, pull_request]
+
+jobs:
+  black-check:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.12'
+
+      - name: Install black
+        run: pip install black
+
+      - name: Run black in check mode
+        run: black --check --diff .
