+++ backend/src/db/repositories/__init__.py
@@ -1,0 +1,1 @@
+from .users import user_repo
