+++ backend/src/services/auth.py
@@ -1,24 +1,17 @@
 from typing import Optional
 from datetime import datetime, timedelta, timezone
 
-from fastapi import Depends, HTTPException, status
 import jwt
-from jwt.exceptions import PyJWTError
+from jwt import ExpiredSignatureError, InvalidTokenError
+from fastapi import HTTPException, status
 from passlib.context import CryptContext
-from sqlalchemy import or_
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.future import select
-from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 
 from src.core.config import get_env
 from src.db.models import User
-from src.db.session import get_async_session
-from src.schemas.user import UserCreate
 from src.schemas.auth import TokenPair
 
 config = get_env()
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-http_bearer = HTTPBearer()
 
 
 def hash_password(password: str) -> str:
@@ -62,11 +55,11 @@ def create_token_pair(user: User) -> TokenPair:
         "sub": str(user.id),
         "username": user.username,
         "email": user.email,
-        "type": "access"
+        "type": "access",
     }
     refresh_token_payload = {
         "sub": str(user.id),
-        "type": "refresh"
+        "type": "refresh",
     }
     return TokenPair(
         access_token=create_access_token(access_token_payload),
@@ -88,63 +81,14 @@ def decode_access_token(token: str):
                 detail="Invalid token type",
             )
         return payload
-    except PyJWTError:
-        return None
-
-
-async def register_user(
-    user_create: UserCreate,
-    session: AsyncSession,
-) -> User:
-    existing_user = await session.scalar(
-        select(User).where(
-            (User.email == user_create.email) | (User.username == user_create.username)
-        )
-    )
-    if existing_user:
-        raise ValueError("User with this email or username already exists")
-
-    user = User(
-        username=user_create.username,
-        email=user_create.email,
-        hashed_password=hash_password(user_create.password),
-    )
-    session.add(user)
-    await session.commit()
-    await session.refresh(user)
-    return create_token_pair(user)
-
-
-async def authenticate_user(
-    identifier: str,
-    password: str,
-    session: AsyncSession,
-) -> User | None:
-    user = await session.scalar(
-        select(User).where(
-            or_(
-                User.username == identifier,
-                User.email == identifier,
-            )
-        )
-    )
-
-    if not user or not verify_password(password, user.hashed_password):
-        raise ValueError("Invalid credentials")
-
-    return create_token_pair(user)
-
-
-async def get_current_user(
-    session: AsyncSession = Depends(get_async_session),
-    credentials: HTTPAuthorizationCredentials = Depends(http_bearer),
-):
-    payload = decode_access_token(credentials.credentials)
-    user_id = payload.get("sub", None)
-    if user_id is None:
+    except ExpiredSignatureError:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="No user found",
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Token has expired",
         )
 
-    return await session.scalar(select(User).where(User.id == user_id))
+    except InvalidTokenError:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid token",
+        )
